+++ hidden/Extensions/Notification.Name+Extension.swift
@@ -13,5 +13,4 @@ extension Notification.Name {
     
     static let prefsChanged = Notification.Name("prefsChanged")
     static let alwayHideToggle = Notification.Name("alwayHideToggle")
-    static let enginePreferenceChanged = Notification.Name("enginePreferenceChanged")
 }